Refactor Iac project command line validations (#3273)
* fix: integration test case on wall time handler

* Refactor Iac project command line validations

* add project reslover/validator parameters

* resolve the test cases

* Add unit tests

* small refactor on the cmd options and the resolver

* tiny refactor, add comments to explain the validations

* rename to plugin_additional_options

* Move get_iac_plugin inside IacProjectResolver

* trigger tests
