+++ samcli/commands/local/invoke/cli.py
@@ -6,12 +6,13 @@ import logging
 import click
 
 from samcli.cli.main import pass_context, common_options as cli_framework_options, aws_creds_options, print_cmdline_args
-from samcli.commands._utils.iac_validations import iac_options_validation
-from samcli.commands._utils.options import project_type_click_option, cdk_click_options
+from samcli.commands._utils.iac_project_validator import IacProjectValidator
+from samcli.commands._utils.options import plugin_additional_options
 from samcli.commands.local.cli_common.options import invoke_common_options, local_common_options
 from samcli.commands.local.lib.exceptions import InvalidIntermediateImageError
 from samcli.lib.iac.interface import IacPlugin, Project
-from samcli.lib.iac.utils.helpers import inject_iac_plugin
+
+from samcli.lib.iac.utils.iac_project_resolver import IacProjectResolver
 from samcli.lib.telemetry.metric import track_command
 from samcli.cli.cli_config_file import configuration_option, TomlProvider
 from samcli.lib.utils.version_checker import check_newer_version
@@ -46,15 +47,12 @@ STDIN_FILE_NAME = "-"
     "is not specified, no event is assumed. Pass in the value '-' to input JSON via stdin",
 )
 @click.option("--no-event", is_flag=True, default=True, help="DEPRECATED: By default no event is assumed.", hidden=True)
-@project_type_click_option(include_build=True)
 @invoke_common_options
 @local_common_options
 @cli_framework_options
 @aws_creds_options
 @click.argument("function_logical_id", required=False)
-@cdk_click_options
-@inject_iac_plugin(with_build=True)
-@iac_options_validation(require_stack=False)
+@plugin_additional_options
 @pass_context
 @track_command  # pylint: disable=R0914
 @check_newer_version
@@ -82,15 +80,19 @@ def cli(
     config_env,
     container_host,
     container_host_interface,
+    project_type,
     cdk_app,
     cdk_context,
-    project_type: str,
-    iac: IacPlugin,
-    project: Project,
 ):
     """
     `sam local invoke` command entry point
     """
+    click_ctx = click.get_current_context()
+    project_type, iac, project = IacProjectResolver(click_ctx).resolve_project(
+        with_build=True, include_build_folder=True
+    )
+    project_validator = IacProjectValidator(click_ctx, project)
+    project_validator.iac_options_validation(require_stack=False)
     # All logic must be implemented in the ``do_cli`` method. This helps with easy unit testing
     do_cli(
         ctx,
